+++ CHANGELOG.rst
@@ -1,6 +1,26 @@
+0.2.0
+~~~~~
+07.07.2021
+   * Updated docs, cleanup github actions;
+   * Moved ``pytest-sugar`` and ``pytest-deadfixtures`` to dev dependencies;
+   * Added ``ipdb`` and ``iptyon`` to dev dependencies;
+   * Added ``update.sh`` script, to update all the dependencies to last version;
+   * Updating pytz (2020.5 -> 2021.1)
+   * Installing types-pytz (2021.1.0)
+   * Updating flake8 (3.8.4 -> 3.9.2)
+   * Updating flake8-plugin-utils (1.3.1 -> 1.3.2)
+   * Updating flake8-annotations (2.5.0 -> 2.6.2)
+   * Updating flake8-comprehensions (3.3.1 -> 3.5.0)
+   * Updating flake8-multiline-containers (0.0.17 -> 0.0.18)
+   * Updating flake8-pytest-style (1.3.0 -> 1.5.0)
+   * Updating flake8-simplify (0.12.0 -> 0.14.1)
+   * Updating pytest (6.2.1 -> 6.2.4)
+   * Updating pytest-asyncio (0.14.0 -> 0.15.1)
+   * Updating pytest-cov (2.11.1 -> 2.12.1)
+
 0.1.5
 ~~~~~
-25.01.2020
+25.01.2021
    * Add some flake8 plugins and bandit;
 
 0.1.4